+++ tcod/bsp.py
@@ -27,7 +27,8 @@ Example::
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Iterator
+from collections.abc import Iterator
+from typing import TYPE_CHECKING, Any
 
 from typing_extensions import deprecated